Refactoring OnboardingScreen
