+++ lib/ui/screens/res/constants.dart
@@ -53,3 +53,9 @@ const String textNotSelected = 'Не выбрано';
 const String textEnterNamePlace = 'Введите название места';
 const String textEnterDescPlace = 'Заполните описание';
 const String textEnterText = 'введите текст';
+const String textOnboardingScreenFirstTitle = 'Добро пожаловать\nв Путеводитель';
+const String textOnboardingScreenFirstDesc = 'Ищи новые локации и сохраняй\nсамые любимые.';
+const String textOnboardingScreenSecondTitle = 'Построй маршрут\nи отправляйся в путь';
+const String textOnboardingScreenSecondDesc = 'Достигай цели максимально\nбыстро и комфортно.';
+const String textOnboardingScreenThirdTitle = 'Добавляй места,\nкоторые нашёл сам';
+const String textOnboardingScreenThirdDesc = 'Делись самыми интересными\nи помоги нам стать лучше!';
